1. signature refactoring for loadAndExecuteStrategy 2. added implementation for LoadAndExecuteStrategyForm.java

+++ Assignments/PDP-Assignment-10/src/main/java/virtualgambling/controller/Features.java
@@ -82,7 +82,7 @@ public interface Features {
   );
 
   boolean loadAndExecuteStrategy(String portfolioName, String filePath, BigDecimal amountToInvest
-          , int commissionPercentage);
+          , double commissionPercentage);
 
   boolean loadPortfolio(String filePath);
